Feat/sql generation (#19)
* (feat): add poetry

* (feat): add poetry

* (feat): add uuid and encryption

* (feat): add uuid and encryption

* Development (#14)

* add env example and launch.json example

* Feat/sql generation (#11)

* (feat): add poetry

* (feat): add poetry

* (feat): add uuid and encryption

* (feat): add uuid and encryption

* Feat/sql generation (#11) (#12)

* (feat): add poetry

* (feat): add poetry

* (feat): add uuid and encryption

* (feat): add uuid and encryption

* fix/delete env.copy

---------


* (fix): fix scanned status

* (fix): fix scanned status

* (fix): fix scanned status

* (feat): add sql generation, execution, llm, and business glossary module

* (feat): add sql generation, execution, llm, and business glossary module

---------
