--- .env.copy
@@ -1,17 +1,0 @@
-APP_NAME=KAI API
-APP_DESCRIPTION="KAI stands for Knowledge Agent for Intelligence query. This project brings the Gen AI component to be able to embedded into the database so that it can perform analytics and document searches with natural language."
-APP_VERSION=1.0.0
-APP_ENVIRONMENT=LOCAL
-
-APP_HOST=0.0.0.0
-APP_PORT=8000
-APP_ENABLE_HOT_RELOAD=0
-
-TYPESENSE_API_KEY = kai_typesense
-TYPESENSE_HOST = localhost
-TYPESENSE_PORT = 8108
-TYPESENSE_PROTOCOL = HTTP
-TYPESENSE_TIMEOUT = 2
-
-#Encryption key for storing DB connection data in Mongo
-ENCRYPT_KEY = f0KVMZHZPgdMStBmVIn2XD049e6Mun7ZEDhf1W7MRnw=
