+++ .env.example
@@ -7,11 +7,27 @@ APP_HOST=0.0.0.0
 APP_PORT=8000
 APP_ENABLE_HOT_RELOAD=0
 
-TYPESENSE_API_KEY = # Give your typesense API KEY
+TYPESENSE_API_KEY = kai_typesense
 TYPESENSE_HOST = localhost
 TYPESENSE_PORT = 8108
 TYPESENSE_PROTOCOL = HTTP
 TYPESENSE_TIMEOUT = 2
 
+OPENAI_API_KEY = 
+OLLAMA_API_BASE= 
+HUGGINGFACEHUB_API_TOKEN= 
+# While using Azure, mention the embedding model here. If you are using OpenAI, use the "text-embedding-3-large"
+
+# use different tools you can set the "AGENT_MAX_ITERATIONS" env variable. By default it is set to 20 iterations.
+AGENT_MAX_ITERATIONS = 20
+#timeout in seconds for the engine to return a response. Defaults to 150 seconds
+DH_ENGINE_TIMEOUT = 150
+#timeout for SQL execution, our agents execute the SQL query to recover from errors, this is the timeout for that execution. Defaults to 60 seconds
+SQL_EXECUTION_TIMEOUT = 60
+#The upper limit on number of rows returned from the query engine (equivalent to using LIMIT N in PostgreSQL/MySQL/SQlite). Defauls to 50
+UPPER_LIMIT_QUERY_RETURN_ROWS = 50
 #Encryption key for storing DB connection data in Typesense
 ENCRYPT_KEY = f0KVMZHZPgdMStBmVIn2XD049e6Mun7ZEDhf1W7MRnw=
+
+# While using Azure, mention the embedding model here. If you are using OpenAI, use the "text-embedding-3-large"
+EMBEDDING_MODEL = "text-embedding-3-large"
