+++ app/api/requests.py
@@ -1,5 +1,6 @@
-from pydantic import BaseModel, field_validator
 from typing import Optional
+
+from pydantic import BaseModel, field_validator
 from sql_metadata import Parser
 
 from app.modules.sql_generation.models import LLMConfig
@@ -42,6 +43,13 @@ class PromptRequest(BaseModel):
     text: str
     db_connection_id: str
     schemas: list[str] | None = None
+    metadata: dict | None = None
+
+
+class BusinessGlossaryRequest(BaseModel):
+    metric: str
+    alias: list[str] | None = None
+    sql: str
     metadata: dict | None = None
 
 
@@ -92,6 +100,13 @@ class NLGenerationsSQLGenerationRequest(NLGenerationRequest):
 
 class PromptSQLGenerationNLGenerationRequest(NLGenerationRequest):
     sql_generation: PromptSQLGenerationRequest
+
+
+class UpdateBusinessGlossaryRequest(BaseModel):
+    metric: str | None = None
+    alias: list[str] | None = None
+    sql: str | None = None
+    metadata: dict | None = None
 
 
 class UpdateMetadataRequest(BaseModel):
