+++ app/api/responses.py
@@ -27,6 +27,13 @@ class PromptResponse(BaseResponse):
     schemas: list[str] | None
 
 
+class BusinessGlossaryResponse(BaseResponse):
+    db_connection_id: str
+    metric: str
+    alias: list[str] | None
+    sql: str
+
+
 class InstructionResponse(BaseResponse):
     db_connection_id: str
     condition: str
@@ -34,8 +41,9 @@ class InstructionResponse(BaseResponse):
     condition_embedding: list[float] | None = None
     is_default: bool
 
+
 class SQLGenerationResponse(BaseResponse):
-    id:str
+    id: str
     prompt_id: str
     status: str
     llm_config: LLMConfig | None
