+++ app/modules/sql_generation/models/__init__.py
@@ -19,14 +19,14 @@ class IntermediateStep(BaseModel):
 class SQLGeneration(BaseModel):
     id: str | None = None
     prompt_id: str
-    llm_config: LLMConfig | None 
+    llm_config: LLMConfig | None
     evaluate: bool = False
     intermediate_steps: list[IntermediateStep] | None = None
     sql: str | None = None
     status: str = "INVALID"
-    tokens_used: int | None 
+    tokens_used: int | None = 0
     confidence_score: float | None = None
-    completed_at: datetime | None = None
+    completed_at: str | None = None
     error: str | None = None
-    created_at: datetime = Field(default_factory=datetime.now)
+    created_at: str = Field(default_factory=lambda: datetime.now().isoformat())
     metadata: dict | None = None
