--- app/utils/encrypt.py
@@ -1,19 +1,0 @@
-from cryptography.fernet import Fernet
-
-from app.server.config import Settings
-
-
-class FernetEncrypt:
-    def __init__(self):
-        settings = Settings()
-        self.fernet_key = Fernet(settings.require("ENCRYPT_KEY"))
-
-    def encrypt(self, input: str) -> str:
-        if not input:
-            return ""
-        return self.fernet_key.encrypt(input.encode()).decode("utf-8")
-
-    def decrypt(self, input: str) -> str:
-        if input == "":
-            return ""
-        return self.fernet_key.decrypt(input).decode("utf-8")
