+++ app/utils/sql_evaluator/simple_evaluator.py
@@ -122,7 +122,7 @@ class SimpleEvaluator(Evaluator):
         db_scan = repository.get_all_tables_by_db(
             {
                 "db_connection_id": str(database_connection.id),
-                "status": TableDescriptionStatus.SCANNED.value,
+                "sync_status": TableDescriptionStatus.SCANNED.value,
             }
         )
         self.llm = self.model.get_model(
