+++ app/utils/sql_tools/__init__.py
@@ -1,0 +1,4 @@
+def replace_unprocessable_characters(text: str) -> str:
+    """Replace unprocessable characters with a space."""
+    text = text.strip()
+    return text.replace(r"\_", "_")
